Update: 配置信息使用类存储

+++ python/utils/PathUtils.py
@@ -7,14 +7,20 @@ from Config import Config
 class PathUtils(object):
 
     @staticmethod
-    def rebuild_dir(dir_path: str):
+    def rebuild_dir(dir_path: str, skip=False) -> bool:
         """
         删除并重新建立文件夹
         :param dir_path: 文件夹路径
+        :param skip: 文件夹存在时是否跳过
+        :return 是否新建了文件夹
         """
         if os.path.exists(dir_path):
-            shutil.rmtree(dir_path)
+            if skip:
+                return False
+            else:
+                shutil.rmtree(dir_path)
         os.mkdir(dir_path)
+        return True
 
     @staticmethod
     def join_path(*path_args):
